% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_cols.R
\name{coffee_cols}
\alias{coffee_cols}
\title{Return one of the ten coffee colors}
\usage{
coffee_cols(...)
}
\arguments{
\item{...}{Columns desired, either the bare names, or the numbers
associated with the colors.}
}
\value{
The hex code of the desired color.
}
\description{
Coffee House Ponders theme
1 - mocha_splatter
2 - coffee_grounds
3 - espresso_cream
4 - latte
5 - flower_sheet_ligth

Cappucino theme
6 - hot_milk
7 - cream
8 - cream_and_coffee
9 - beans
10 - espresso
}
\details{
Part of the Rstatscoffeegeeks essentials. Return the
color code associated by the name and number.
}
\examples{
coffee_col(1)
coffee_col(mocha_splatter)
coffee_col(hot_milk, espresso)
coffee_col(1, 5, 6)
coffee_col(espresso, 4, 6)

ggplot2::ggplot(mtcars, aes(drat, mpg)) +
  geom_point(aes(col = as.factor(cyl))) +
  scale_color_manual(values = coffee_cols(beans, latte, espresso))
}
