% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffeegeeks.R
\docType{package}
\name{coffeegeeks}
\alias{coffeegeeks}
\alias{coffeegeeks-package}
\title{Coffeegeeks}
\description{
A package full of coffee goodness. including but not
limited to coffee themes for ggplot.
}
