% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_coffee.R
\name{coffee_pal}
\alias{coffee_pal}
\alias{scale_colour_coffee}
\alias{scale_color_coffee}
\alias{scale_fill_coffee}
\alias{scale_gradient_coffee}
\title{A set of coffee colors}
\usage{
coffee_pal()

scale_colour_coffee(...)

scale_color_coffee(...)

scale_fill_coffee(...)

scale_gradient_coffee(..., space = "Lab", na.value = "grey50",
  guide = "colourbar")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}

\item{na.value}{Colour to use for missing values}
}
\description{
A set of coffee colors
}
