% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flavours.R
\name{extract_flavour}
\alias{extract_flavour}
\title{Extract flavours from sentance}
\usage{
extract_flavour(charvec, description = FALSE)
}
\arguments{
\item{charvec}{a sentance like: "The coffee has a smooth soft creamy taste"}

\item{description}{FALSE or TRUE if you want the description of those words}
}
\description{
This function extracts the flavours from a sentance and
returns them with or without descriptions.
The default returns just the words that are tastes, and
if you set description to TRUE a data frame with word and explanation is returned.
}
\examples{
extract_flavour("the nippy brew had a oily after taste with turpsy tones")
}
