% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flavours.R
\name{get_flavour_descriptions}
\alias{get_flavour_descriptions}
\title{words to describe coffee taste}
\source{
\url{http://articles.submityourarticle.com/words-used-to-describe-the-taste-of-coffee-55923}
}
\usage{
get_flavour_descriptions()
}
\description{
The following flavours are used to describe coffee.
}
