% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_taste.R
\name{help_my_coffee_}
\alias{help_my_coffee_}
\title{Help my coffee ...}
\source{
\url{"http://drinks.seriouseats.com/2013/07/home-coffee-brewing-troubleshooting-grind-size-brew-time.html"}

\url{"http://www.jimseven.com/2006/07/31/5-reasons-your-coffee-tastes-bad/"}

\url{"https://www.roastycoffee.com/reasons-your-coffee-tastes-bad/"}
}
\usage{
help_my_coffee_(tastes = NULL, is = NULL)
}
\description{
Helps you diagnose trouble in coffee paradise.
See examples for instructions.
}
\details{
You can add multiple things wrong with your coffee
}
\examples{
help_my_coffee_(is = "cold")
help_my_coffee_(tastes = "sour")
help_my_coffee_(tastes = "bitter")
help_my_coffee_(is = "gone")
help_my_coffee_(tastes = c("sour", "strong"))
}
